Progress: Continue migrating the project.
Start brining in the enumerations and types.
There is still a lot of inconsistency in the type delete functions to deal with.
Many of the "delete" functions may end up being replaced with a "resize" method.

This removes the "--init" parameter.
I figure I might as well simplify the design a little and have the init and controller be more separate.