Progress: Continue migrating the project, notably refactoring out controller_global_t.
A lot of the changes made it possible to remove th controller_global_t.
Restructure the code to do this.

Most of the compile errors are fixed.

I still need to go over the logic and fix things.
I still need to copy over some functions.
I still need to move the print functions around.
I still need to do runtime tests once all of the above is completed.