Progress: Continue migrating the project, where the code now compiles without error.
The migration is not complete.
This simply has all of the compilation problems addressed.

There are still several print functions that need to be migrated.
There are probably a lot of out of date comments, which will be fixed as discovered.

I need to walk through each and every function, from a start to finish perspective, and make sure that all the logic is still valid.
I have good reason to believe that there are a lot of things to correct.