Cleanup: Add time based version in each of the documentation and specification files where it is missing.
I am just setting the current date rather than trying to figure out the date of last change.

I also noticed that the documents/time.txt file has an incorrect license specified.
The original files from the 0.6 and 0.7 branches of the FLL project already show the open-standard-license-1.0-or-later as the license.
Make sure the documents/time.txt correctly identifies the license as open-standard-license-1.0-or-later to avoid confusion.